use the nicer workspace consumer
